% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_generator.R
\name{graph_generator}
\alias{graph_generator}
\title{Title}
\usage{
graph_generator(
  n = 200,
  d = 50,
  graph = "SBM",
  v = NULL,
  u = NULL,
  p = 10,
  prob = NULL,
  vis = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{verbose}{}
}
\description{
Title
}
