% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moonshoot.R
\name{moonshoot}
\alias{moonshoot}
\title{Title}
\usage{
moonshoot(
  X,
  SigmaHat,
  ThetaHat,
  q,
  numB = 1000,
  V0,
  possibleSet,
  prescreen = T
)
}
\arguments{
\item{prescreen}{}
}
\description{
Title
}
