% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_bootstrap_normalize_test.R
\name{set_bootstrap_normalize_test}
\alias{set_bootstrap_normalize_test}
\title{Title}
\usage{
set_bootstrap_normalize_test(data, Theta.hat, num.B)
}
\arguments{
\item{num.B}{}
}
\description{
Title
}
